+++ storage/storage.go
@@ -2,5 +2,5 @@ package storage
 
 type Storager interface {
 	GetAny(key string) (any, error)
-	StoreAny(key string, value any, ttl uint) error
+	StoreAny(key string, value any, ttl int64) error
 }
